Added getStartDate endpoints as a controller.
