+++ api/src/controllers/getStartDate.controller.ts
@@ -1,0 +1,24 @@
+import * as express from "express";
+import * as notangles from "../notangles/index";
+import * as freerooms from "../freerooms/index";
+
+/**
+ * Controller for getting the start date of the latest term for
+ * Freerooms and Notangles.
+ * @param req
+ * @param res
+ */
+
+const getStartDateByRouteParams = (req: express.Request, res: express.Response) => {
+    try {
+        const client = req.params.client;
+        if (client === "notangles") {
+            res.send(notangles.getStartDate(req, res));
+        } else if (client === "freerooms") {
+            res.send(freerooms.getStartDate(req, res));
+        }
+    } catch (e) {
+        res.status(400).send("Error");
+    }
+};
+export { getStartDateByRouteParams };
